system-design/python/src/auth/requirements.txt
astroid==2.9.3
blinker==1.7.0
click==8.0.3
dill==0.3.7
Flask==2.0.2
Flask-MySQLdb==0.2.0
importlib-metadata==6.8.0
isort==5.10.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.1.0
parso==0.8.3
platformdirs==2.4.1
PyJWT==2.3.0
pylint==2.12.2
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.8.0
Werkzeug==2.0.2
zipp==3.17.0
